Renamed service and middleware files
